EnemyManager has been converted to an object. (no monobehaviour)
GUI Manager has been checked in and is linked with mana invasion

Apparently changing the binding to code base will remove all those problems on the GUI.

Fahad, for the spawn function, the instantiate doesnt work anymore without monobehaviour.
Tell me if you cannot work with this style. (Without monobehaviour for the manager)

Signed-off-by: John Tai <zapyre_1@hotmail.com>